backend/requirements.txt
fastapi
uvicorn[standard]
lightgbm
numpy
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
alembic
